Set isSelected on StripLayoutTab as part of starting state
This ensures that keyboard focus states remain consistent when the selected tab is scrolled off screen.

Screencast after: https://screencast.googleplex.com/cast/NDkxNDYxODA3MjgyNTg1NnwxMzBlNWIxZS0xNQ

Screencast before: https://screencast.googleplex.com/cast/NjM2MDI5Njk1MTExOTg3Mnw0ODQyNWZkNi1hNA

Fixed: 419072631
Test: 
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6570950
Commit-Queue: Jenna Himawan <jhimawan@google.com>
Cr-Commit-Position: refs/heads/main@{#1463509}

+++ chrome/android/java/src/org/chromium/chrome/browser/compositor/overlays/strip/StripLayoutHelper.java
@@ -1063,6 +1063,9 @@ public class StripLayoutHelper
             computeAndUpdateTabWidth(
                     /* animate= */ false, /* deferAnimations= */ false, /* closedTab= */ null);
         }
+        if (getSelectedTabId() != Tab.INVALID_TAB_ID) {
+            tabSelected(LayoutManagerImpl.time(), getSelectedTabId(), Tab.INVALID_TAB_ID);
+        }
     }
 
     /** Called to notify that the tab state has been initialized. */
@@ -1091,6 +1094,9 @@ public class StripLayoutHelper
         // Recreate the StripLayoutTabs from the TabModel, now that all of the real Tabs have been
         // restored. This will reuse valid tabs, discard invalid tabs, and correct tab orders.
         rebuildStripTabs(false, false);
+        if (getSelectedTabId() != Tab.INVALID_TAB_ID) {
+            tabSelected(LayoutManagerImpl.time(), getSelectedTabId(), Tab.INVALID_TAB_ID);
+        }
     }
 
     /**
@@ -1381,7 +1387,9 @@ public class StripLayoutHelper
             updateCloseButtons();
 
             Tab tab = getTabById(id);
-            if (tab != null && mTabGroupModelFilter.getTabGroupCollapsed(tab.getRootId())) {
+            if (tab != null
+                    && mTabGroupModelFilter != null
+                    && mTabGroupModelFilter.getTabGroupCollapsed(tab.getRootId())) {
                 mTabGroupModelFilter.deleteTabGroupCollapsed(tab.getRootId());
             }
 
@@ -1394,7 +1402,9 @@ public class StripLayoutHelper
             mUpdateHost.requestUpdate();
 
             setAccessibilityDescription(stripTab, getTabById(id));
-            setAccessibilityDescription(findTabById(prevId), getTabById(prevId));
+            if (prevId != Tab.INVALID_TAB_ID) {
+                setAccessibilityDescription(findTabById(prevId), getTabById(prevId));
+            }
         }
         StripLayoutTab previouslyFocusedTab = findTabById(prevId);
         if (previouslyFocusedTab != null) previouslyFocusedTab.setIsSelected(false);
